[TIME-SLOT-WEEK-VIEW-COMP][VM|DS] Refactor to remove unused typings
